Made inv2 the primary inventory. Added some code for stacking. Added stacking to the inventory UI (Thanks @Tim). added some temp debug items
