+++ Assets/PreFabs/Items/Food/InvItem-Apple.asset
@@ -13,9 +13,11 @@ MonoBehaviour:
   m_Name: InvItem-Apple
   m_EditorClassIdentifier: 
   InvData:
+    Id: 15100ddc-f577-4131-b486-68c49f9d1ee4
     ItemName: Apple
     ItemIcon: {fileID: 21300000, guid: 776bfbc244f69b44bbd9b4e43e966fe4, type: 3}
     ItemPrefab: {fileID: 4922043878047182297, guid: 9d54c94494a3d944a83e090425251335, type: 3}
+    MaxStackSize: 2
     ItemType: 0
     AnimationSet: 
     Description:
