+++ Assets/PreFabs/Items/Food/InvItem-Bread.asset
@@ -13,9 +13,11 @@ MonoBehaviour:
   m_Name: InvItem-Bread
   m_EditorClassIdentifier: 
   InvData:
+    Id: 8f3cb7fd-be4a-465d-b33e-29ce164ae2c6
     ItemName: Bread
     ItemIcon: {fileID: 21300000, guid: 5d0a717872335c946bc141cf41ed62c7, type: 3}
     ItemPrefab: {fileID: 9049046711321312459, guid: 9edc1a83723edea4184177ac3a65fb80, type: 3}
+    MaxStackSize: 2
     ItemType: 0
     AnimationSet: 
     Description:
