+++ Assets/PreFabs/Items/Food/InvItem-Paddo.asset
@@ -13,9 +13,11 @@ MonoBehaviour:
   m_Name: InvItem-Paddo
   m_EditorClassIdentifier: 
   InvData:
+    Id: 22c50495-0251-4ba7-b65b-ea59ee58ada8
     ItemName: Paddo
     ItemIcon: {fileID: 21300000, guid: c087fbc27dd89c342acaf4e96c6bd9b3, type: 3}
     ItemPrefab: {fileID: 3158172204144373338, guid: 681270435d9059b469d15da79dd75c1f, type: 3}
+    MaxStackSize: 2
     ItemType: 0
     AnimationSet: 
     Description:
