+++ Assets/PreFabs/Items/Food/InvItem-Watermelon.asset
@@ -13,9 +13,11 @@ MonoBehaviour:
   m_Name: InvItem-Watermelon
   m_EditorClassIdentifier: 
   InvData:
+    Id: 91203627-5c9f-4852-bafc-e5d6d4a64dbb
     ItemName: Watermelon
     ItemIcon: {fileID: 21300000, guid: 3c2b078da2d02504aa75a40342871e23, type: 3}
     ItemPrefab: {fileID: 7922604427852444920, guid: 233a46e766431d04e8f6ca2b43677d28, type: 3}
+    MaxStackSize: 2
     ItemType: 0
     AnimationSet: 
     Description:
