+++ Assets/PreFabs/Items/Sword/InvItem-Sword.asset
@@ -11,11 +11,13 @@ MonoBehaviour:
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 64c0e1789899e29448350646d58e3256, type: 3}
   m_Name: InvItem-Sword
-  m_EditorClassIdentifier:
+  m_EditorClassIdentifier: 
   InvData:
+    Id: 4ebd3b27-ff50-42ce-a1da-2cc9c6a0a5bc
     ItemName: Sword
     ItemIcon: {fileID: 21300000, guid: 5bea5df02f4f92e23aadad14997cee84, type: 3}
     ItemPrefab: {fileID: 6458623882520386215, guid: 73f704797938366419d8190a387a5fab, type: 3}
+    MaxStackSize: 1
     ItemType: 1
     AnimationSet: Sword
-    Description:
+    Description:
