+++ Assets/PreFabs/Systems/HUD.prefab
@@ -45,7 +45,7 @@ MonoBehaviour:
   m_Script: {fileID: 19102, guid: 0000000000000000e000000000000000, type: 0}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_PanelSettings: {fileID: 11400000, guid: dc71905af099c6f419e95a4de31fafd7, type: 2}
+  m_PanelSettings: {fileID: 11400000, guid: a415c0cbf408967408be98d18aa2d9aa, type: 2}
   m_ParentUI: {fileID: 0}
   sourceAsset: {fileID: 9197481963319205126, guid: 7aa619a453e8a044a94b48c39c1b33a7, type: 3}
   m_SortingOrder: 0
@@ -63,4 +63,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   HasPlaytestKey: 0
   Health: 1
-  Dash: 1
+  DashStart: -1.75
+  _dashReadyColor: {r: 0, g: 1, b: 0.5, a: 1}
+  _dashUnreadyColor: {r: 0, g: 0.5, b: 1, a: 1}
