--- Assets/Scripts/Controllers/Inventory/Inv2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 41fe7550c8c137052b8a5ad935a1cf61
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
