+++ Assets/UnityDefaultRuntimeTheme 1.tss
@@ -1,0 +1,2 @@
+@import url("unity-theme://default");
+VisualElement {}
