+++ Assets/UnityDefaultRuntimeTheme 1.tss.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 61e3e414ca7864fd49e7f3475819585d
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 12388, guid: 0000000000000000e000000000000000, type: 0}
+  disableValidation: 0
